Fixing Build - moved LogEvent extensions into its own file
